+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - BC breaks (follow `UPGRADE-3.0.md` to upgrade):
   - [#101]: support for concurrent instances of the same flow
   - removed the step field template
+  - renamed property `step` to `stepNumber` and method `getStep` to `getStepNumber` within event classes
 
 [#98]: https://github.com/craue/CraueFormFlowBundle/issues/98
 [#101]: https://github.com/craue/CraueFormFlowBundle/issues/101